case-studies: rewrite case studies for grammar evaluation
As a first step, we compare the grammar that is the result from the
grammar semantics with a grammar created from the set of terms from the
wildcard semantics. A simple language equality and -inclusion check can
then already inform us whether the grammar semantics give more precise
results or not.